Check that session hasn't expired

+++ src/services/delete-account.ts
@@ -1,15 +1,18 @@
 import type { AccountRes } from "../pages/api/account/[id]";
 import ApiRoutes from "../routing/api-routes";
 import { en } from "./context/language-ctx";
-import authHeader from "./headers/auth-header";
+import getAuthHeader from "./headers/auth-header";
 import Notification from "./notifications/notification";
 
 export default async function deleteAccount(id: number): Promise<AccountRes | null> {
+  const authHeader = await getAuthHeader();
+  if (!authHeader) return null;
+
   const notification = new Notification();
   try {
     notification.loading(en ? "Deleting Account..." : "Suppression du compte...");
     const res = await fetch(ApiRoutes.deleteAccount(id), {
-      headers: await authHeader(),
+      headers: authHeader,
       method: "DELETE",
     });
     if (!res.ok) throw await res.text();
